migrate to calypso-products
